requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-cors-headers==3.11.0
djangorestframework==3.12.4
Pillow==9.0.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
